requirements.txt
python-fasthtml>=1.0.0
monsterui>=0.1.0
python-frontmatter>=1.0.0
Markdown>=3.5.0
uvicorn>=0.25.0
